+++ src/contexts/changeConfiguration/managedApiKey.ts
@@ -1,5 +1,6 @@
 import { verifyApiKey } from '@app/utilities/openai'
 import { waitFor } from '@app/utilities/node'
+import { ConfigurationService } from '@app/services'
 
 export class ManagedApiKey {
   private static instance: ManagedApiKey
@@ -18,6 +19,7 @@ export class ManagedApiKey {
     this._isQueued = true
     await waitFor(500, () => false)
     await verifyApiKey()
+    ConfigurationService.LogConfigurationService()
     this._isQueued = false
   }
 }