+++ src/quickPicks/quickPickSetupAzureOpenai.ts
@@ -9,9 +9,8 @@
 
 import { QuickPickItem, CancellationToken, ExtensionContext, Uri } from 'vscode'
 import { ConfigurationService } from '@app/services'
-import { azureListDeployments } from '@app/utilities/openai'
+import { listModelsAzureOpenAI, ModelCapabiliy } from '@app/utilities/openai'
 import { SecretStorageService, MultiStepInput } from '@app/utilities/vscode'
-import { createInfoNotification } from '@app/utilities/node'
 
 /**
  * This function sets up a quick pick menu for configuring the OpenAI service provider.
@@ -27,7 +26,8 @@ export async function quickPickSetupAzureOpenai(
     totalSteps: number
     openaiBaseUrl: string
     openaiApiKey: string
-    openaiModel: QuickPickItem
+    quickPickInferenceModel: QuickPickItem
+    quickPickEmbeddingModel: QuickPickItem
   }
 
   async function collectInputs() {
@@ -51,7 +51,7 @@ export async function quickPickSetupAzureOpenai(
     state.openaiBaseUrl = await input.showInputBox({
       title,
       step: 1,
-      totalSteps: 3,
+      totalSteps: 4,
       ignoreFocusOut: true,
       value:
         typeof state.openaiBaseUrl === 'string'
@@ -81,7 +81,7 @@ export async function quickPickSetupAzureOpenai(
     state.openaiApiKey = await input.showInputBox({
       title,
       step: 2,
-      totalSteps: 3,
+      totalSteps: 4,
       ignoreFocusOut: true,
       value: typeof state.openaiApiKey === 'string' ? state.openaiApiKey : '',
       prompt: 'Enter you openai.com Api-Key',
@@ -89,7 +89,7 @@ export async function quickPickSetupAzureOpenai(
       validate: validateAzureOpenaiApiKey,
       shouldResume: shouldResume,
     })
-    return (input: MultiStepInput) => selectOpenaiModel(input, state)
+    return (input: MultiStepInput) => selectChatCompletionModel(input, state)
   }
 
   /**
@@ -97,26 +97,59 @@ export async function quickPickSetupAzureOpenai(
    * @param input - The multi-step input object.
    * @param state - The current state of the application.
    */
-  async function selectOpenaiModel(
+  async function selectChatCompletionModel(
     input: MultiStepInput,
     state: Partial<State>
   ) {
     const models = await getAvailableModels(
       state.openaiApiKey!,
       state.openaiBaseUrl!,
+      ModelCapabiliy.ChatCompletion,
       undefined /* TODO: token */
     )
     // Display quick pick menu for selecting an OpenAI model and update application's state accordingly.
     // Return void since this is not used elsewhere in the code.
-    state.openaiModel = await input.showQuickPick({
+    state.quickPickInferenceModel = await input.showQuickPick({
       title,
       step: 3,
-      totalSteps: 3,
+      totalSteps: 4,
       ignoreFocusOut: true,
       placeholder:
-        'Selected DeploymentModel (if empty, no valid chat completion models found)',
+        'Selected Chat Completion DeploymentModel (if empty, no valid chat completion models found)',
       items: models,
-      activeItem: state.openaiModel,
+      activeItem: state.quickPickInferenceModel,
+      shouldResume: shouldResume,
+    })
+
+    return (input: MultiStepInput) => selectEmbeddingModel(input, state)
+  }
+
+  /**
+   * This function displays a quick pick menu for selecting an OpenAI model and updates the application's state accordingly.
+   * @param input - The multi-step input object.
+   * @param state - The current state of the application.
+   */
+  async function selectEmbeddingModel(
+    input: MultiStepInput,
+    state: Partial<State>
+  ) {
+    const models = await getAvailableModels(
+      state.openaiApiKey!,
+      state.openaiBaseUrl!,
+      ModelCapabiliy.Embedding,
+      undefined /* TODO: token */
+    )
+    // Display quick pick menu for selecting an OpenAI model and update application's state accordingly.
+    // Return void since this is not used elsewhere in the code.
+    state.quickPickEmbeddingModel = await input.showQuickPick({
+      title,
+      step: 4,
+      totalSteps: 4,
+      ignoreFocusOut: true,
+      placeholder:
+        'Selected Embedding DeploymentModel (if empty, no valid chat completion models found)',
+      items: models,
+      activeItem: state.quickPickEmbeddingModel,
       shouldResume: shouldResume,
     })
   }
@@ -153,11 +186,13 @@ export async function quickPickSetupAzureOpenai(
   async function getAvailableModels(
     openapiAPIKey: string,
     openapiBaseUrl: string,
+    modelCapabiliy: ModelCapabiliy,
     token?: CancellationToken
   ): Promise<QuickPickItem[]> {
-    const chatCompletionModels = await azureListDeployments(
+    const chatCompletionModels = await listModelsAzureOpenAI(
       openapiAPIKey,
-      openapiBaseUrl
+      openapiBaseUrl,
+      modelCapabiliy
     )
 
     const quickPickItems: QuickPickItem[] = []
@@ -168,9 +203,6 @@ export async function quickPickSetupAzureOpenai(
       })
     })
     return quickPickItems
-
-    // Map each returned label into a QuickPickItem object with label property set as label value returned by API call.
-    // return chatCompletionModels.map((label) => ({ label }))
   }
 
   function shouldResume() {
@@ -182,13 +214,17 @@ export async function quickPickSetupAzureOpenai(
 
   //Start openai.com configuration processes
   const state = await collectInputs()
+
+  const inferenceModel = state.quickPickInferenceModel.description as string
+  const inferenceDeployment = state.quickPickInferenceModel.label
+  const embeddingModel = state.quickPickEmbeddingModel.description as string
+  const embeddingDeployment = state.quickPickEmbeddingModel.label
+
   ConfigurationService.instance.serviceProvider = 'Azure-OpenAI'
   ConfigurationService.instance.baseUrl = state.openaiBaseUrl
-  ConfigurationService.instance.azureDeployment = state.openaiModel.label
-  ConfigurationService.instance.defaultModel = state.openaiModel
-    .description as string
+  ConfigurationService.instance.azureDeployment = inferenceDeployment
+  ConfigurationService.instance.defaultModel = inferenceModel
+  ConfigurationService.instance.embeddingsDeployment = embeddingDeployment
+  ConfigurationService.instance.embeddingModel = embeddingModel
   SecretStorageService.instance.setAuthApiKey(state.openaiApiKey)
-  createInfoNotification(
-    `ServiceProvider (Azure): ${state.openaiBaseUrl} ${state.openaiModel.label}`
-  )
 }