+++ utilities/scripts/clean.js
@@ -1,7 +1,13 @@
 const fs = require('fs')
 const path = require('path')
 
-const folders = ['../../dist', '../../out', '../../node_modules']
+const folders = [
+  '../../dist',
+  '../../out',
+  '../../node_modules',
+  '../../webview-ui/conversationsWebview/node_modules',
+  '../../webview-ui/messageWebview/node_modules',
+]
 
 folders.forEach((folder) => {
   const dir = path.join(__dirname, folder)